% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valencia_script.R
\name{Valencia}
\alias{Valencia}
\title{Valencia CST Classification}
\usage{
Valencia(input_df)
}
\arguments{
\item{input_df}{A dataframe of bacterial count data that is in the proper format for Valencia analysis.
The first and second column will need to be titled sampleID, read_counts. All following column names must be formatted to fit
Valencia reference database naming convention.}
}
\value{
Returns a data frame with CST classifications and confidence scores for each sample
}
\description{
Valencia will analyze vaginal microbiome data to obtain CST classification
}
